Add draft of task widget separation
